Revert "making the api great again"
This reverts commit 9af5e0addd425f1c3f944b66f103828a1e10ec6f.

+++ app/Http/Controllers/Api/Ingest/AccountArchiveController.php
@@ -31,10 +31,16 @@ class AccountArchiveController extends Controller
      */
     public function index(Request $request, Account $account)
     {
-        // TODO: setup guards
-        $archives = Account::with('archives')->first()->archives()->with('metadata');
+        // todo: the selection needs to be modified when roles and groups gets fully implemented
+        if($account->owner_id !== auth()->user()->id) {
+            abort( response()->json([ 'error' => 404, 'message' => 'Invalid account' ], 404 ) );
+        }
+
+        $archives = $account->archives();
+
         $limit = $request->limit ? $request->limit : env('DEFAULT_ENTRIES_PER_PAGE');
         return ArchiveResource::collection( $archives->paginate( $limit ) );
+
     }
 
     /**
@@ -43,9 +49,14 @@ class AccountArchiveController extends Controller
      * @param \Illuminate\Http\Request $request
      * @param File $file
      * @return void
-,    */
+     */
     public function store(Request $request, Account $account)
     {
+        // todo: the selection needs to be modified when roles and groups gets fully implemented
+        if($account->owner_id !== auth()->user()->id) {
+            abort( response()->json([ 'error' => 404, 'message' => 'Invalid account' ], 404 ) );
+        }
+
         $requestData = $this->validateRequest($request);
         $archive = new Archive(
             array_merge($requestData, [
@@ -71,6 +82,11 @@ class AccountArchiveController extends Controller
      */
     public function show(Account $account, Archive $archive)
     {
+        // todo: the selection needs to be modified when roles and groups gets fully implemented
+        if($account->owner_id !== auth()->user()->id) {
+            abort( response()->json([ 'error' => 404, 'message' => 'Invalid account' ], 404 ) );
+        }
+
         return response()->json([ "data" => new ArchiveResource($archive)]);
     }
 
@@ -83,9 +99,15 @@ class AccountArchiveController extends Controller
      */
     public function update(Request $request, Account $account, Archive $archive)
     {
-        $requestData = $this->validateRequest( $request );
-        $archive->update( $requestData );
-        return new ArchiveResource( $archive );
+        // todo: the selection needs to be modified when roles and groups gets fully implemented
+        if($account->owner_id !== auth()->user()->id) {
+            abort( response()->json([ 'error' => 404, 'message' => 'Invalid account' ], 404 ) );
+        }
+
+        $requestData = $this->validateRequest($request);
+        $archive->update($requestData);
+        $archive->save();
+        return response()->json([ "data" => new ArchiveResource($archive)]);
     }
 
     /**
@@ -97,6 +119,11 @@ class AccountArchiveController extends Controller
      */
     public function destroy(Account $account, Archive $archive)
     {
+        // todo: the selection needs to be modified when roles and groups gets fully implemented
+        if($account->owner_id !== auth()->user()->id) {
+            abort( response()->json([ 'error' => 404, 'message' => 'Invalid account' ], 404 ) );
+        }
+
         $archive->delete();
         return response( "", 204);
     }
